"Refactored RCD application code: added date input modal, updated embed titles, and modified button logic and responses."

+++ bot/regular_commands/rcd_aplication.py
@@ -1,3 +1,6 @@
+from datetime import datetime
+import re
+
 import discord
 from discord.ui import Modal, InputText, View, button, select
 from loguru import logger
@@ -12,8 +15,67 @@ from variables import VETERAN_ROLE
 member_list: list = []
 channel_last_message: list[discord.Message] = []
 rcd_channel: list[discord.TextChannel] = []
-embed: discord.Embed = final_rcd_list_embed()
-last_message_rcd_list: list[discord.Message] = []
+rcd_date_list: list[str] = []
+embed: list[discord.Embed] = []
+last_message_to_finish: dict[int: discord.Message] = {}
+
+
+class RcdDate(Modal):
+    """
+    Модальное окно для ввода даты РЧД.
+
+    Attributes:
+    ----------
+        date: str
+            Дата проведения РЧД.
+    """
+    def __init__(self):
+        super().__init__(title='Введи дату проведения РЧД', timeout=None)
+
+        self.add_item(
+            InputText(
+                style=discord.InputTextStyle.multiline,
+                label='Укажи дату в формате ДД.ММ',
+                placeholder='ДД.ММ',
+                max_length=5
+            )
+        )
+
+    async def callback(self, interaction: discord.Interaction):
+        date_str: str = str(self.children[0].value)
+        date_pattern = r'^([0-2][0-9]|3[0-1])[.,/](0[1-9]|1[0-2])$'
+        date_match = re.match(date_pattern, date_str)
+
+        if not date_match:
+            return await interaction.respond(
+                'Неправильный формат даты. Пожалуйста, используйте формат ДД.ММ',
+                ephemeral=True,
+                delete_after=5
+            )
+
+        try:
+            day, month = map(int, date_match.groups())
+            current_year = datetime.now().year
+            rcd_date = datetime(
+                year=current_year,
+                month=month,
+                day=day)
+            if rcd_date < datetime.now():
+                rcd_date = rcd_date.replace(year=current_year + 1)
+            convert_rcd_date = discord.utils.format_dt(rcd_date, style="D")
+            rcd_date_list.append(convert_rcd_date)
+            embed.append(final_rcd_list_embed(convert_rcd_date))
+            await interaction.respond(embed=rcd_list_embed())
+            await interaction.respond(view=StartRCDButton())
+            await interaction.respond(
+                '_РЧД заявки запущены!_',
+                ephemeral=True,
+                delete_after=2
+            )
+        except Exception as error:
+            logger.error(
+                f'При вводе даты РЧД возникла ошибка {error}'
+            )
 
 
 class RaidChampionDominionApplication(Modal):
@@ -165,8 +227,8 @@ class SelectMemberToRCD(View):
         interaction: discord.Interaction,
         value: str
     ) -> None:
-        embed.fields[self.index].value = value
-        await last_message_rcd_list[0].edit(embed=embed)
+        embed[0].fields[self.index].value = value
+        await last_message_to_finish.get(3).edit(embed=embed[0])
         await interaction.message.delete()
         await interaction.respond(
             f'Добавлено {"✅" if value else "⭕"}',
@@ -214,9 +276,10 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
+        last_message_to_finish[2] = interaction.message
         obj_1: discord.ui.Button = self.children[0]
         obj_2: discord.ui.Button = self.children[1]
-        obj_13: discord.ui.Select = self.children[12]
+        obj_14: discord.ui.Select = self.children[13]
 
         obj_1.label = '⬇️ Список создан ниже ⬇️'
         obj_1.style = discord.ButtonStyle.gray
@@ -225,9 +288,9 @@ class CreateRCDList(View):
         obj_2.style = discord.ButtonStyle.green
         obj_2.disabled = False
 
-        obj_13.disabled = False
+        obj_14.disabled = False
         await interaction.message.edit(view=self)
-        await interaction.respond(embed=embed)
+        await interaction.respond(embed=embed[0])
 
     @button(
         label='Добавить "Воинов"', style=discord.ButtonStyle.gray,
@@ -238,8 +301,8 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        last_message_rcd_list.append(interaction.channel.last_message)
-        await self.update_view_rcd_list(interaction, 'Воины:', 0)
+        last_message_to_finish[3] = interaction.channel.last_message
+        await self.update_view_rcd_list(interaction, 0)
 
     @button(
         label='Добавить "Паладинов"', style=discord.ButtonStyle.gray,
@@ -250,17 +313,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 2:
-            embed.add_field(
-                name='Паладин:',
-                value='',
-                inline=False
-            )
-        obj_4: discord.ui.Button = self.children[3]
-        obj_4.style = discord.ButtonStyle.green
-        obj_4.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=1))
+        await self.update_view_rcd_list(interaction, 1)
 
     @button(
         label='Добавить "Инженеров"', style=discord.ButtonStyle.gray,
@@ -271,17 +324,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 3:
-            embed.add_field(
-                name='Инженеры:',
-                value='',
-                inline=False
-            )
-        obj_5: discord.ui.Button = self.children[4]
-        obj_5.style = discord.ButtonStyle.green
-        obj_5.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=2))
+        await self.update_view_rcd_list(interaction, 2)
 
     @button(
         label='Добавить "Жрецов"', style=discord.ButtonStyle.gray,
@@ -292,17 +335,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 4:
-            embed.add_field(
-                name='Жрецы:',
-                value='',
-                inline=False
-            )
-        obj_6: discord.ui.Button = self.children[5]
-        obj_6.style = discord.ButtonStyle.green
-        obj_6.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=3))
+        await self.update_view_rcd_list(interaction, 3)
 
     @button(
         label='Добавить "Друидов"', style=discord.ButtonStyle.gray,
@@ -313,17 +346,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 5:
-            embed.add_field(
-                name='Шаман:',
-                value='',
-                inline=False
-            )
-        obj_7: discord.ui.Button = self.children[6]
-        obj_7.style = discord.ButtonStyle.green
-        obj_7.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=4))
+        await self.update_view_rcd_list(interaction, 4)
 
     @button(
         label='Добавить "Мистиков"', style=discord.ButtonStyle.gray,
@@ -334,17 +357,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 6:
-            embed.add_field(
-                name='Мистики:',
-                value='',
-                inline=False
-            )
-        obj_8: discord.ui.Button = self.children[7]
-        obj_8.style = discord.ButtonStyle.green
-        obj_8.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=5))
+        await self.update_view_rcd_list(interaction, 5)
 
     @button(
         label='Добавить "Лучников"', style=discord.ButtonStyle.gray,
@@ -355,17 +368,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 7:
-            embed.add_field(
-                name='Лучники:',
-                value='',
-                inline=False
-            )
-        obj_9: discord.ui.Button = self.children[8]
-        obj_9.style = discord.ButtonStyle.green
-        obj_9.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=6))
+        await self.update_view_rcd_list(interaction, 6)
 
     @button(
         label='Добавить "Магов"', style=discord.ButtonStyle.gray,
@@ -376,17 +379,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 8:
-            embed.add_field(
-                name='Маги:',
-                value='',
-                inline=False
-            )
-        obj_10: discord.ui.Button = self.children[9]
-        obj_10.style = discord.ButtonStyle.green
-        obj_10.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=7))
+        await self.update_view_rcd_list(interaction, 7)
 
     @button(
         label='Добавить "Некромантов"', style=discord.ButtonStyle.gray,
@@ -397,17 +390,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 9:
-            embed.add_field(
-                name='Некроманты:',
-                value='',
-                inline=False
-            )
-        obj_11: discord.ui.Button = self.children[10]
-        obj_11.style = discord.ButtonStyle.green
-        obj_11.disabled = False
-        await interaction.message.edit(view=self)
-        await interaction.respond(view=SelectMemberToRCD(index=8))
+        await self.update_view_rcd_list(interaction, 8)
 
     @button(
         label='Добавить "Бардов"', style=discord.ButtonStyle.gray,
@@ -418,7 +401,7 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        await self.update_view_rcd_list(interaction, 'Барды:', 9)
+        await self.update_view_rcd_list(interaction, 9)
 
     @button(
         label='Добавить "Демонологов"', style=discord.ButtonStyle.gray,
@@ -429,13 +412,30 @@ class CreateRCDList(View):
         button: discord.ui.Button,
         interaction: discord.Interaction
     ):
-        if len(embed.fields) < 11:
-            embed.add_field(
+        if len(embed[0].fields) < 11:
+            embed[0].add_field(
                 name='Демоны:',
                 value='',
                 inline=False
             )
         await interaction.respond(view=SelectMemberToRCD(index=10))
+
+    @button(
+        label='Завершить работу со списком РЧД', style=discord.ButtonStyle.red,
+        custom_id='ЗавершитьРЧДСписок', disabled=True
+    )
+    async def stop_callback(
+        self,
+        button: discord.ui.Button,
+        interaction: discord.Interaction
+    ):
+        await interaction.channel.delete_messages([message for message in last_message_to_finish.values()])
+        await interaction.respond(
+            '_Работа со списком РЧД завершена!_',
+            ephemeral=True,
+            delete_after=2
+
+        )
 
     @select(
         select_type=discord.ComponentType.channel_select,
@@ -448,12 +448,18 @@ class CreateRCDList(View):
         self, select: discord.ui.Select, interaction: discord.Interaction
     ):
         channel: discord.TextChannel = select.values[0]
-        await channel.last_message.delete()
-        await channel.send(embed=embed)
+        if 'Список РЧД' in channel.last_message.embeds[0].title:
+            await channel.last_message.edit(embed=embed[0])
+        else:
+            await channel.last_message.delete()
+            await channel.send(embed=embed[0])
+        button: discord.ui.Button = self.children[12]
+        button.disabled = False
+        await interaction.message.edit(view=self)
         await interaction.respond(
             f'_Список РЧД опубликован в канале {channel.mention}_',
             ephemeral=True,
-            delete_after=5
+            delete_after=2
         )
 
     async def update_view_rcd_list(
@@ -461,8 +467,8 @@ class CreateRCDList(View):
             interaction: discord.Interaction,
             index: int,
     ) -> None:
-        if len(embed.fields) < index + 1:
-            embed.add_field(
+        if len(embed[0].fields) < index + 1:
+            embed[0].add_field(
                 name=self.index_class_role.get(index),
                 value='',
                 inline=False
@@ -499,10 +505,11 @@ class StartRCDButton(View):
     async def select_callback(
         self, select: discord.ui.Select, interaction: discord.Interaction
     ):
+        last_message_to_finish[1] = interaction.message
         channel: discord.TextChannel = select.values[0]
         try:
             await channel.send(
-                embed=start_rcd_embed(),
+                embed=start_rcd_embed(rcd_date_list[0]),
                 view=RCDButton(
                     embed=rcd_list_embed()
                 )
